requirements.txt
# H-DCD 项目依赖清单
# PyTorch 需要单独安装，请参考 README.md 中的安装说明

# ============================================
# 核心依赖 (必须安装)
# ============================================

# Mamba 状态空间模型核心依赖
# PyTorch 2.5.0+ 使用最新版本
causal-conv1d>=2.0.0
mamba-ssm>=2.2.0

# PyTorch 2.0-2.1 使用旧版本
# causal-conv1d>=1.4.0,<2.0.0
# mamba-ssm>=1.2.0,<2.0.0

# 数值计算和张量操作
numpy>=1.23.0,<2.0.0
scipy>=1.9.0
einops>=0.7.0

# 数据处理
pandas>=1.5.0
scikit-learn>=1.2.0

# 配置文件
pyyaml>=6.0

# 进度条
tqdm>=4.64.0

# 包管理工具
packaging>=21.0
ninja>=1.11.0

# ============================================
# 可视化工具 (推荐)
# ============================================
matplotlib>=3.6.0
seaborn>=0.12.0
tensorboard>=2.12.0

# ============================================
# 多模态数据处理 (可选)
# ============================================
# 视觉特征提取
opencv-python>=4.7.0

# 音频特征提取
librosa>=0.10.0
soundfile>=0.12.0

# 文本特征提取 (BERT/RoBERTa)
transformers>=4.30.0
tokenizers>=0.13.0

# ============================================
# 开发工具 (可选)
# ============================================
# 代码测试
pytest>=7.2.0
pytest-cov>=4.0.0

# 代码格式化
black>=23.0.0
isort>=5.12.0

# 代码检查
flake8>=6.0.0
pylint>=2.17.0

# 类型检查
mypy>=1.3.0

# Jupyter 支持
ipython>=8.12.0
jupyter>=1.0.0
notebook>=6.5.0

# ============================================
# 分布式训练 (可选)
# ============================================
# 多 GPU 训练支持
# torch.distributed (PyTorch 自带)

# Weights & Biases 实验跟踪
wandb>=0.15.0

# ============================================
# 版本兼容性说明
# ============================================
# Python: 3.9 - 3.12 (推荐 3.10 或 3.11)
# PyTorch: 2.0.0 - 2.5.1 (推荐 2.5.0+，性能提升 20-30%)
# CUDA: 11.8 或 12.1+ (推荐 12.1+)
#
# 安装 PyTorch 命令示例：
#
# 最新版本 (CUDA 12.1, 推荐):
# conda install pytorch torchvision torchaudio pytorch-cuda=12.1 -c pytorch -c nvidia
# 或
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# PyTorch 2.5.0 (CUDA 12.1):
# pip install torch==2.5.0 torchvision==0.20.0 torchaudio==2.5.0 --index-url https://download.pytorch.org/whl/cu121
#
# PyTorch 2.1.0 (CUDA 11.8, 稳定版):
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu118
#
# CPU 版本:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
